+++ Lab.DTO/Masters/Interfaces/IMstCasePaper.cs
@@ -9,7 +9,7 @@ namespace Lab.DTO.Masters.Interfaces
 {
     public interface IMstCasePaper
     {
-        int Create(DTOCasePaper _objDtoCasePaper);
+        Int64 Create(DTOCasePaper _objDtoCasePaper);
         Task<string> GetLastPatientIdForDate(string datePart);
     }
 }
